# yahriacad/docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: YahriaCad API
  description: |
    API REST du backend Go de YahriaCad, application de CAO électronique :
    gestion de projets, import de fichiers CAO (KiCad, Eagle, échange natif),
    placement et routage assistés par intelligence artificielle, vérification
    DRC/ERC et exports de fabrication (Gerber, BOM, STEP).

    Conventions générales :
      * Toutes les coordonnées et dimensions sont exprimées en **millimètres**,
        origine en coin haut-gauche de la carte, indice de couche 0 = F.Cu.
      * Les champs JSON sont en **snake_case** (contrat partagé `pcb.d.ts`).
      * Les jobs asynchrones (routage, optimisation) renvoient `202 JobStarted`
        et publient leur progression sur le WebSocket `/ws/v1/progress` (§3
        des contrats) : messages `{"type":"progress",...}` serveur -> client,
        `{"type":"subscribe","project_id":"…"}` et `{"type":"ping"}` client ->
        serveur, `{"type":"pong"}` en retour. Ping applicatif toutes les 30 s.
      * Le serveur démarre même si le moteur IA est injoignable : les routes
        IA renvoient alors `502` avec le code d'erreur `ai_unreachable`.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Développement local
...
paths:
...
  /api/v1/projects/{id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: Détail d'un projet
      description: Retourne les métadonnées du projet (le contenu design passe par /layout et /import).
      operationId: getProject
      responses:
        '200':
          description: Projet trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Projet introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Projects]
      summary: Mise à jour des métadonnées
      description: Renomme le projet et/ou modifie sa description. Les champs omis sont conservés.
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  example: Carte capteur 60x40 rev B
                description:
                  type: string
                  example: Deuxième itération, ajout d'un condensateur de découplage
            example:
              name: Carte capteur 60x40 rev B
              description: Deuxième itération
      responses:
        '200':
          description: Projet mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Projet introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Projects]
      summary: Suppression d'un projet
      description: 'Supprime le projet et son contenu. Idempotent : supprimer un projet inconnu renvoie 204.'
      operationId: deleteProject
      responses:
        '204':
          description: Projet supprimé (ou déjà absent)
        '404':
          description: Projet introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/projects/{id}/drc:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [Vérification]
      summary: Design Rule Check
      description: |
        Vérifie les règles de conception sur le layout courant : clearances
        piste/piste, piste/pad, piste/via, via/via, cuivre/bord, largeur de
        piste, diamètre/perçage/anneau annulaire des vias. Codes de violation :
        `DRC_CLEARANCE`, `DRC_TRACK_WIDTH`, `DRC_VIA_DIAMETER`, `DRC_DRILL`,
        `DRC_ANNULAR`, `DRC_EDGE_CLEARANCE`.
      operationId: runDRC
      responses:
        '200':
          description: Rapport DRC (passed = false si au moins une violation de gravité error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DRCResult'
        '404':
          description: Projet introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{id}/erc:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [Vérification]
      summary: Electrical Rule Check
      description: |
        Vérifie la cohérence électrique du schéma : références dupliquées,
        nets à une seule broche, nets sans nom, empreinte manquante,
        composant non connecté. Codes : `ERC_DUPLICATE_REF`,
        `ERC_SINGLE_PIN_NET`, `ERC_EMPTY_NET_NAME`, `ERC_MISSING_FOOTPRINT`,
        `ERC_UNCONNECTED_COMPONENT`.
      operationId: runERC
      responses:
        '200':
          description: Rapport ERC
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERCResult'
        '404':
          description: Projet introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      description: Identifiant unique du projet (UUID).
      schema:
        type: string
        format: uuid
      example: 7d1e5c2a-9b3f-4c11-a6d8-2f0e8b91c4d7
...
  schemas:
    Project:
      type: object
      description: Métadonnées d'un projet YahriaCad.
      required: [id, name, description, layer_count, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: 7d1e5c2a-9b3f-4c11-a6d8-2f0e8b91c4d7
        name:
          type: string
          example: Carte capteur 60x40
        description:
          type: string
          example: Prototype deux couches, connecteur J1 et microcontrôleur U1
        layer_count:
          type: integer
          minimum: 1
          maximum: 34
          example: 2
        status:
          type: string
          enum: [active, archived]
          example: active
        created_at:
          type: string
          format: date-time
          example: '2025-01-15T08:30:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-01-15T10:12:45Z'
...
    DRCViolation:
      type: object
      description: Violation d'une règle de conception, localisée sur la carte.
      required: [code, severity, message, x, y, layer, net]
      properties:
        code:
          type: string
          enum:
            - DRC_CLEARANCE
            - DRC_TRACK_WIDTH
            - DRC_VIA_DIAMETER
            - DRC_DRILL
            - DRC_ANNULAR
            - DRC_EDGE_CLEARANCE
          example: DRC_CLEARANCE
        severity:
          type: string
          enum: [error, warning]
          example: error
        message:
          type: string
          example: "Clearance 0,175 mm < 0,45 mm requis entre les nets VCC et GND"
        x:
          type: number
          example: 18.25
        y:
          type: number
          example: 10.3
        layer:
          type: integer
          example: 0
        net:
          type: string
          example: VCC
    DRCResult:
      type: object
      description: Rapport DRC complet.
      required: [passed, checked_rules, violations, duration_ms]
      properties:
        passed:
          type: boolean
          example: false
        checked_rules:
          type: integer
          example: 6
        violations:
          type: array
          items:
            $ref: '#/components/schemas/DRCViolation'
        duration_ms:
          type: integer
          example: 14
    ERCViolation:
      type: object
      description: Anomalie électrique détectée sur le schéma.
      required: [code, severity, message, component_ref]
      properties:
        code:
          type: string
          enum:
            - ERC_DUPLICATE_REF
            - ERC_SINGLE_PIN_NET
            - ERC_EMPTY_NET_NAME
            - ERC_MISSING_FOOTPRINT
            - ERC_UNCONNECTED_COMPONENT
          example: ERC_SINGLE_PIN_NET
        severity:
          type: string
          enum: [error, warning]
          example: warning
        message:
          type: string
          example: "le net \"AUX\" n'est relié qu'à une seule broche"
        component_ref:
          type: string
          example: R1
    ERCResult:
      type: object
      description: Rapport ERC complet.
      required: [passed, violations]
      properties:
        passed:
          type: boolean
          example: true
        violations:
          type: array
          items:
            $ref: '#/components/schemas/ERCViolation'
...
    ErrorResponse:
      type: object
      description: Enveloppe d'erreur uniforme de l'API.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [not_found, invalid, conflict, ai_unreachable, internal]
              example: not_found
            message:
              type: string
              example: "projet : introuvable"
      example:
        error:
          code: not_found
          message: "projet : introuvable"

# ==============================================================
# WebSocket /ws/v1/progress (hors OpenAPI paths, voir contracts.md §3) :
#
#   URL        : ws://localhost:8080/ws/v1/progress?project_id=<id>
#   Client     : {"type":"subscribe","project_id":"..."} ou {"type":"ping"}
#   Serveur    : {"type":"progress","job_id":"...","project_id":"...",
#                 "stage":"route","current_net":"GND","percent":42,
#                 "message":"...","done":false,"error":""}
#                {"type":"pong"}
#   Filtrage   : par project_id (query OU message subscribe).
#   Keepalive  : ping applicatif toutes les 30 s.
# ==============================================================